+++ src/test-utils/test-wrapper.jsx
@@ -1,17 +1,17 @@
 // @flow
 import React, { type Node } from "react";
 // $FlowFixMe: there is no createGenerateClassName export in @material-ui/core/styles
-import { createGenerateClassName } from "@material-ui/styles";
-/* eslint-disable import/no-extraneous-dependencies */
-import { JssProvider } from "react-jss";
 import { StyleWrapper } from "../style-wrapper";
+import { StylesProvider, createGenerateClassName } from "@material-ui/styles";
 
 type Props = {
   children: Node<*>,
 };
 
+// TODO(nsawas): Figure out why this new method of disabling global styles
+// isnt exactly working. For now we disable all styles with `disableGeneration` prop below.
 const generateClassName = createGenerateClassName({
-  dangerouslyUseGlobalCSS: true,
+  disableGlobal: true,
 });
 
 class MockResizeObserver {
@@ -24,8 +24,11 @@ window.ResizeObserver = MockResizeObserver;
 export const TestWrapper = (props: Props) => {
   const { children } = props;
   return (
-    <JssProvider generateClassName={generateClassName}>
+    <StylesProvider
+      disableGeneration
+      generateClassName={generateClassName}
+    >
       <StyleWrapper>{children}</StyleWrapper>
-    </JssProvider>
+    </StylesProvider>
   );
 };