+++ CHANGELOG.md
@@ -48,6 +48,7 @@
 42. Make LargeSimpleTable sortable.
 43. Add `autosizeHeight` prop to `LargeSimpleTable`.
 44. Add optional `subheader` prop to SignInPage.
+45. `AvatarIcon` now accepts `children`.
 
 ## 1.0.0 - 2019-06-17